refactor(policy): move the safeguard resolver into the policy package
