+++ typescript/packages/core/src/workspace/executor/fanout.ts
@@ -19,7 +19,7 @@ import { PathSpec } from '../../types.ts'
 import type { MountEntry } from '../mount/mount.ts'
 import type { MountRegistry } from '../mount/registry.ts'
 import { ExecutionNode } from '../types.ts'
-import { resolveAcrossMounts } from '../../commands/safeguard.ts'
+import { resolveAcrossMounts } from './policy/safeguard.ts'
 import { applyFindActions } from './find_action_dispatch.ts'
 import { respellOne } from '../../utils/path.ts'
 import { rstripSlash } from '../../utils/slash.ts'
